added listener for email verification, reworked controller so it would return only pageable requests, renamed some packages for better understanding
