+++ .gitignore
@@ -35,5 +35,3 @@ build/
 
 ### scripts ###
 /scripts/
-
-.mvn
