--- src/main/java/com/softserve/message_sender/service/MailService.java
@@ -1,10 +1,0 @@
-package com.softserve.message_sender.service;
-
-import com.softserve.message_sender.model.MessageHandler;
-
-public interface MailService {
-
-    void sendReceiptMessage(MessageHandler messageHandler);
-
-    //TODO email verification
-}
