+++ src/main/java/com/softserve/messaging/model/MessageHandler.java
@@ -1,20 +1,20 @@
-package com.softserve.message_sender.model;
+package com.softserve.messaging.model;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.Transient;
 import java.time.ZonedDateTime;
 import java.util.UUID;
 
 @Entity
 @Data
 @NoArgsConstructor
+@AllArgsConstructor
 public class MessageHandler {
 
     @Id
@@ -22,26 +22,11 @@ public class MessageHandler {
     private Long id;
 
     private UUID userUUID;
-    private String messageBody;
+    private String firstName;
+    private String lastName;
+    private String mail;
+    private double cost;
+    private String token;
     private ZonedDateTime zonedDateTime;
 
-    @Transient
-    private String firstName;
-    @Transient
-    private String lastName;
-    @Transient
-    private String mail;
-    @Transient
-    private double cost;
-    @Transient
-    private String token;
-
-    public static final String TITLE = "Electronic receipt";
-
-    @Override
-    public String toString() {
-        return String.format("Hello, %s. Thanks for a ride. We hope you've enjoyed it.%nTotal cost: %.2f $", getFirstName(), getCost());
-    }
-
-    //TODO util class for static strings?
 }
