+++ src/main/java/com/softserve/messaging/service/MailService.java
@@ -1,0 +1,8 @@
+package com.softserve.messaging.service;
+
+import com.softserve.messaging.dto.MailTemplateDto;
+
+public interface MailService {
+
+    void sendMessage(MailTemplateDto mailTemplateDto);
+}
