+++ src/main/java/com/softserve/messaging/service/MailServiceImpl.java
@@ -1,6 +1,7 @@
-package com.softserve.message_sender.service;
+package com.softserve.messaging.service;
 
-import com.softserve.message_sender.model.MessageHandler;
+import com.softserve.messaging.dto.MailTemplateDto;
+import com.softserve.messaging.model.MessageHandler;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.mail.javamail.JavaMailSender;
@@ -17,17 +18,18 @@ public class MailServiceImpl implements MailService {
 
     private final JavaMailSender javaMailSender;
 
-    public void sendReceiptMessage(MessageHandler messageHandler) {
+    public void sendMessage(MailTemplateDto mailTemplateDto) {
         MimeMessage message = javaMailSender.createMimeMessage();
 
         try {
             MimeMessageHelper helper = new MimeMessageHelper(message, true);
 
-            helper.setTo(messageHandler.getMail());
-            helper.setSubject(MessageHandler.TITLE);
-            helper.setText(messageHandler.toString());
+            helper.setTo(mailTemplateDto.getMail());
+            helper.setSubject(mailTemplateDto.getTopic());
+            helper.setText(mailTemplateDto.getBody());
 
             javaMailSender.send(message);
+
             log.info("Mail was successfully sent");
         } catch (MessagingException e) {
             log.error("Mail wasn't sent, exception: {}", e.getMessage());
