+++ src/test/java/com/softserve/messaging/MessageSenderApplicationTests.java
@@ -1,4 +1,4 @@
-package com.softserve.message_sender;
+package com.softserve.messaging;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
